Add editable user note to posts
Introduced a UserNoteForm component and integrated it into post creation and viewing. Users can now add a note when creating a post. The development of editing the note after the creation of the post is in progress. Updated database handler and hooks to support saving and updating user notes. Upgraded appwrite and related dependencies.